+++ Core/CMakeLists.txt
@@ -56,6 +56,7 @@ set(SRCS
   HLE/sceNp.cpp
   HLE/sceP3da.cpp
   HLE/sceMp4.cpp
+  HLE/sceMp3.cpp
   HLE/scePauth.cpp
   HLE/scePower.cpp
   HLE/scePsmf.cpp
