+++ Core/Core.vcxproj
@@ -211,6 +211,7 @@
     <ClCompile Include="HLE\scePspNpDrm_user.cpp" />
     <ClCompile Include="HLE\sceP3da.cpp" />
     <ClCompile Include="HLE\sceMp4.cpp" />
+    <ClCompile Include="HLE\sceMp3.cpp" />
     <ClCompile Include="HLE\sceGameUpdate.cpp" />
     <ClCompile Include="HLE\sceRtc.cpp" />
     <ClCompile Include="HLE\sceSas.cpp" />
@@ -383,6 +384,7 @@
     <ClInclude Include="HLE\scePspNpDrm_user.h" />
     <ClInclude Include="HLE\sceP3da.h" />
     <ClInclude Include="HLE\sceMp4.h" />
+    <ClInclude Include="HLE\sceMp3.h" />
     <ClInclude Include="HLE\sceGameUpdate.h" />
     <ClInclude Include="HLE\sceRtc.h" />
     <ClInclude Include="HLE\sceSas.h" />
