+++ Core/HLE/sceMpeg.h
@@ -96,4 +96,3 @@ void __MpegDoState(PointerWrap &p);
 void __MpegShutdown();
 
 void Register_sceMpeg();
-void Register_sceMp3();
